UPDATE: better NPE handling in DynamicRecordBuffer.close()
git-svn-id: https://svn.javlin.eu/svn/cloveretl.opensource/branches/release-3-5-0@15448 a09ad3ba-1a0f-0410-b1b9-c67202f10d70

+++ cloveretl.engine/src/org/jetel/data/DynamicRecordBuffer.java
@@ -565,11 +565,15 @@ public class DynamicRecordBuffer {
 			try {
 				fullFileBuffers = null;
 		        emptyFileBuffers = null;
-				tempFileChannel.close();
-			} finally {
-		        if (!tempFile.delete()) {
-		        	log.warn("Failed to delete temp file: " + tempFile.getAbsolutePath());
+		        if (tempFileChannel != null) {
+		        	tempFileChannel.close();
 		        }
+			} finally {
+				if (tempFile != null) {
+			        if (!tempFile.delete()) {
+			        	log.warn("Failed to delete temp file: " + tempFile.getAbsolutePath());
+			        }
+				}
 			}
 		}
